Exercice 4/backend/src/etc/requirements.txt
flask
psycopg2-binary
requests
pysocks
